requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
Pillow
tqdm
albumentations>=1.4.24
fastapi>=0.128.0
numpy>=1.26.0
opencv-python-headless>=4.11.0
openai>=1.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
scikit-learn>=1.7.0
torch>=2.10.0
torchvision>=0.25.0
uvicorn[standard]>=0.40.0

